requirements.txt
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
spotipy==2.24.0
urllib3==2.2.3
Werkzeug==3.1.3
